Don't throw if _filename doesn't exist yet
This would all be much simpler if we were on .NET Core 3.0 or later,
but for .NET Framework we have to do a complicated dance to replicate
what would be a simple `mv tmpfile.xml realfile.xml` on Linux.

+++ SIL.BuildTasks/MakeWixForDirTree/IdToGuidDatabase.cs
@@ -242,10 +242,17 @@ namespace SIL.BuildTasks.MakeWixForDirTree
 				}
 				writer.WriteEndElement(); // end InstallerMetadata
 			}
-			File.Replace(tempFilename, _filename, backupFilename);
-			// If that didn't throw an exception, it's now safe to delete the backup file
-			// Catch and suppress any errors caused by deleting the backup file; if that ever happens it should not fail builds
-			try { File.Delete(backupFilename); } catch { }
+			if (File.Exists(_filename))
+			{
+				File.Replace(tempFilename, _filename, backupFilename);
+				// If that didn't throw an exception, it's now safe to delete the backup file
+				// Catch and suppress any errors caused by deleting the backup file; if that ever happens it should not fail builds
+				try { File.Delete(backupFilename); } catch { }
+			}
+			else
+			{
+				File.Move(tempFilename, _filename);
+			}
 		}
 
 		#endregion
